requirements.txt
pyside6>=6.0.0
python-pptx>=0.6.21
Pillow>=10.0.0
requests>=2.31.0
replicate>=0.22.0
packaging>=23.2
deep-translator>=1.11.4
pyinstaller>=6.10.0
cryptography>=41.0.1
google-genai>=0.6.0
PyPDF2>=3.0.0